+++ app/src/main/java/com/example/playlistmaker/mvvm/library/ui/di/CreatePlaylistVMModule.kt
@@ -1,6 +1,6 @@
 package com.example.playlistmaker.mvvm.library.ui.di
 
-import com.example.playlistmaker.mvvm.library.ui.fragment.createrplaylist.CreatePlaylistViewModel
+import com.example.playlistmaker.mvvm.library.ui.view_model.createrplaylist.CreatePlaylistViewModel
 import org.koin.core.module.dsl.viewModel
 import org.koin.dsl.module
