+++ app/src/main/java/com/example/playlistmaker/mvvm/library/ui/di/PlaylistVMModule.kt
@@ -7,6 +7,6 @@ import org.koin.dsl.module
 val playlistVMModule = module {
 
     viewModel {
-        PlaylistViewModel()
+        PlaylistViewModel(get())
     }
 }
