--- app/src/main/java/com/example/playlistmaker/mvvm/library/ui/models/Playlist.kt
@@ -1,7 +1,0 @@
-package com.example.playlistmaker.mvvm.library.ui.models
-
-data class Playlist(
-    val name: String,
-    val description: String?,
-    val coverImageUrl: String?,
-)
